Automatically calc drone bandwidth for Gila and Rattlesnake without 'special' drone modes. This also enables the use of smaller drones on interceptors, frigates and the like, even with the Gila, etc and it will deploy as many drones as bandwidth allows . Gila, 2 mediums or 4 lights
Rattlesnake, 2 heavies, 5 mediums or 5 lights
added changing the drone mode if the appropriate drone type is not available but a smaller drone type is. If you lose all your heavies it will start using mediums and then lights).
added scaling UP to whatever the smallest drone is you have available if we are trying to kill a small target (ceptor/frigate, etc) and we are out of light drones... (medium, heavy, sentries and then fighters))
if their are other ships with reduced drone bandwidth they can easily be added later with minimal effort.
previously we only used small drones to kill targets within 20k if they are NPCs, we will now do the same for Hostile (already aggressed) players in small ships